requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0 